Fix new tests being in wrong file
